test(tui/settings): split float-precision tests into float_tests.rs
Keep both files under the 400-LOC CI cap. Pure move, no behaviour change.

Refs #785
